+++ src/Plugin.php
@@ -6,20 +6,13 @@ use Carbon\Carbon;
 use CraftCms\Cms\Element\Events\SetRoute;
 use CraftCms\Cms\Plugin\Plugin as BasePlugin;
 use CraftCms\Cms\Support\Facades\Twig;
-use CraftCms\Cms\View\Events\SiteTemplateRootsResolving;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\View;
-use Illuminate\Support\Facades\Vite;
 use Override;
 use wsydney76\craft6blade\Listeners\HandleSetRoute;
 use wsydney76\craft6blade\Support\BladeDirectives;
-use wsydney76\craft6blade\Support\BladeIfs;
-use wsydney76\craft6blade\Support\CraftMacros;
 use wsydney76\craft6blade\Twig\Extensions\BladeTwigExtension;
 use function app;
-use function dd;
-use function public_path;
-use function request;
 
 
 class Plugin extends BasePlugin