+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## 2026-06-22
+- Added the `asRelativeTime` function to convert timestamps to relative time formats (e.g., "5 minutes ago").
+
 ## 2026-06-20
 
 Initial commit.